Encapsulated class properties, added validators and factory+setter functions

+++ Domain/Shared/Entities/ITreeEntity.cs
@@ -2,6 +2,6 @@
 
 public interface ITreeEntity
 {
-    public int? ParentId { get; }
+    public long? ParentId { get; }
     public int Level { get; }
 }
